Tg invest analyz/Moex-app/requirements.txt
apimoex==1.3.0
Flask==2.1.0
numpy==1.20.3
pandas==2.0.3
psycopg2==2.9.6
python-dotenv==0.19.2
requests==2.28.2
